SWTB-1 code cleanup
