--- schrottverwaltung/src/main/java/de/fhdortmund/schrottverwaltung/HelloWorldController.java
@@ -1,15 +1,0 @@
-package de.fhdortmund.schrottverwaltung;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/hello")
-public class HelloWorldController {
-
-    @GetMapping("/world")
-    public String helloWorld() {
-        return "Hello World!";
-    }
-}
