+++ schrottverwaltung/src/main/java/de/fhdortmund/schrottverwaltung/mitarbeiter/CreateUserDto.java
@@ -1,4 +1,4 @@
-package de.fhdortmund.schrottverwaltung.security;
+package de.fhdortmund.schrottverwaltung.mitarbeiter;
 
 public record CreateUserDto(String firstname, String lastname, String email, String password) {
 }
