+++ schrottverwaltung/src/main/java/de/fhdortmund/schrottverwaltung/mitarbeiter/MitarbeiterController.java
@@ -1,0 +1,16 @@
+package de.fhdortmund.schrottverwaltung.mitarbeiter;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/user")
+@RequiredArgsConstructor
+public class MitarbeiterController {
+    private final MitarbeiterService mitarbeiterService;
+
+    @PostMapping("/register")
+    private void register(@RequestBody CreateUserDto credentials){
+        mitarbeiterService.register(credentials);
+    }
+}
