--- schrottverwaltung/src/main/java/de/fhdortmund/schrottverwaltung/security/AuthDto.java
@@ -1,4 +1,0 @@
-package de.fhdortmund.schrottverwaltung.security;
-
-public record AuthDto(String username, String password) {
-}
